pyproject.toml
[project]
name = "jira-testplan-bot"
version = "0.1.0"
description = "Generate structured QA test plans from Jira tickets using AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "fastapi>=0.123.9",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pyyaml>=6.0",
    "mcp>=1.1.2",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "greenlet>=3.1.1",
]

[project.scripts]
testplan = "src.cli.main:app"
testplan-mcp = "src.mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
